requirements.txt
# Python 3.10.8 - 2023 Compatible Versions (canonical)
numpy==1.24.3
opencv-contrib-python==4.8.0.76
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2
mediapipe==0.10.21

Flask==2.2.2
Flask-SocketIO==5.3.3
Flask-SSE==1.0.0

firebase==4.0.1
firebase-admin==6.1.0
Pyrebase==3.0.27
python-dotenv==1.0.1

customtkinter==5.1.3
playsound==1.2.2

requests==2.28.2
Pillow==9.5.0

